requirements.txt
-e ../shared/python
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
sqlalchemy==2.0.39
asyncpg==0.30.0
alembic==1.15.1
redis==5.2.1
httpx==0.28.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
authlib==1.5.1
itsdangerous==2.2.0
email-validator==2.2.0
